requirements.txt
APScheduler==3.6.3
cachetools==4.2.2
certifi==2022.6.15
charset-normalizer==2.1.0
cycler==0.11.0
et-xmlfile==1.1.0
fonttools==4.33.3
idna==3.3
kiwisolver==1.4.3
matplotlib==3.5.2
numpy==1.23.0
openai==0.20.0
openpyxl==3.0.10
packaging==21.3
pandas==1.4.2
pandas-stubs==1.4.2.220626
Pillow==9.2.0
pyparsing==3.0.9
python-dateutil==2.8.2
python-telegram-bot==13.13
pytz==2022.1
pytz-deprecation-shim==0.1.0.post0
requests==2.28.1
six==1.16.0
tornado==6.1
tqdm==4.64.0
typing_extensions==4.3.0
tzdata==2022.1
tzlocal==4.2
urllib3==1.26.9
scikit-learn

